refactor(backend): extract model crud module
Move the model functions out of crud/__init__.py into crud/model.py,
following the split already done for source, task, tag and others.
Consumers import the module directly; nothing is re-exported.

Drops test_model_connection, a stub that only ever returned True and had
no callers -- the /models/{id}/test endpoint uses ModelConnectionService
instead.

routers/services.py already aliased the whole crud package as
model_crud; it now aliases the submodule under the same name.
