+++ apps/backend/src/rhesis/backend/app/crud/__init__.py
@@ -1366,148 +1366,6 @@ def get_type_lookup_by_name_and_value(
     )
 
 
-# Model CRUD
-def get_model(
-    db: Session, model_id: uuid.UUID, organization_id: str = None, user_id: str = None
-) -> Optional[models.Model]:
-    """Get a specific model by ID with its related objects and organization filtering"""
-    query = (
-        db.query(models.Model)
-        .options(include(models.Model.provider_type))
-        .filter(models.Model.id == model_id)
-    )
-
-    # Apply organization filtering (SECURITY CRITICAL)
-    if organization_id:
-        from uuid import UUID as UUIDType
-
-        query = query.filter(models.Model.organization_id == UUIDType(organization_id))
-
-    return query.first()
-
-
-def get_models(
-    db: Session,
-    skip: int = 0,
-    limit: int = 10,
-    sort_by: str = "created_at",
-    sort_order: str = "desc",
-    filter: str | None = None,
-    organization_id: str = None,
-    user_id: str = None,
-) -> List[models.Model]:
-    """Get all models with their related objects"""
-    return get_items_detail(
-        db,
-        models.Model,
-        skip,
-        limit,
-        sort_by,
-        sort_order,
-        filter,
-        organization_id=organization_id,
-        user_id=user_id,
-    )
-
-
-def create_model(
-    db: Session, model: schemas.ModelCreate, organization_id: str = None, user_id: str = None
-) -> models.Model:
-    """Create a new model."""
-    return create_item(db, models.Model, model, organization_id, user_id)
-
-
-def update_model(
-    db: Session,
-    model_id: uuid.UUID,
-    model: schemas.ModelUpdate,
-    organization_id: str = None,
-    user_id: str = None,
-) -> Optional[models.Model]:
-    """Update a model."""
-    # First check if the model is protected
-    existing_model = get_model(db, model_id, organization_id)
-    if existing_model and getattr(existing_model, "is_protected", False):
-        # For protected models, only allow updating certain fields
-        # (tags, comments, status, owner, assignee)
-        # Block updates to core model configuration properties
-        protected_fields = {
-            "name",
-            "model_name",
-            "provider_type_id",
-            "key",
-            "endpoint",
-            "is_protected",
-            "icon",
-        }
-
-        # Convert model to dict and check if any protected fields are being updated
-        update_data = (
-            model.model_dump(exclude_unset=True)
-            if hasattr(model, "model_dump")
-            else model.dict(exclude_unset=True)
-        )
-
-        # Check if user is trying to change any protected fields to a different value
-        attempted_protected_updates = []
-        for field in protected_fields:
-            if field in update_data:
-                existing_value = getattr(existing_model, field)
-                new_value = update_data[field]
-                # Only flag as error if the value is actually changing
-                if existing_value != new_value:
-                    attempted_protected_updates.append(field)
-
-        if attempted_protected_updates:
-            fields_str = ", ".join(attempted_protected_updates)
-            raise ValueError(
-                f"Cannot update protected fields ({fields_str}) on system model. "
-                "Only tags, status, owner, and assignee can be modified."
-            )
-
-    return update_item(db, models.Model, model_id, model, organization_id, user_id)
-
-
-def delete_model(
-    db: Session, model_id: uuid.UUID, organization_id: str, user_id: str
-) -> Optional[models.Model]:
-    """Delete a model (protected models cannot be deleted)"""
-    # First check if the model is protected
-    model = get_model(db, model_id, organization_id)
-    if model and getattr(model, "is_protected", False):
-        raise ValueError("Cannot delete protected system model")
-
-    return delete_item(db, models.Model, model_id, organization_id=organization_id, user_id=user_id)
-
-
-def test_model_connection(db: Session, model_id: uuid.UUID) -> bool:
-    """Test the connection to a model's endpoint
-
-    Args:
-        db: Database session
-        model_id: ID of the model to test
-
-    Returns:
-        bool: True if connection test was successful
-
-    Raises:
-        ValueError: If model not found
-        Exception: If connection test fails
-    """
-    # Get the model
-    model = get_model(db, model_id)
-    if not model:
-        raise ValueError(f"Model with id {model_id} not found")
-
-    try:
-        # Here you would implement the actual connection test logic
-        # This could include making a test request to the model's endpoint
-        # For now, we'll just return True
-        return True
-    except Exception as e:
-        raise Exception(f"Failed to test connection: {str(e)}")
-
-
 # Tool CRUD
 _TOOL_RELATED_FIELDS = (include(models.Tool.tool_provider_type),)
