+++ apps/backend/src/rhesis/backend/app/crud/model.py
@@ -1,0 +1,149 @@
+"""CRUD operations for models -- the LLM/embedding provider configurations users register.
+
+Part of the incremental split of the ``crud`` monolith: ``crud/__init__.py`` still holds
+the bulk of the functions, and per-entity modules like this one take over as the code
+around them is touched.
+
+``update_model`` and ``delete_model`` enforce the ``is_protected`` rules for system models,
+the pre-seeded rows an organization gets on onboarding. A protected model rejects any
+change to its core configuration -- name, model_name, provider_type_id, key, endpoint,
+is_protected, icon -- and rejects deletion outright, while tags, status, owner and assignee
+stay editable so users can still organize it. Both raise ``ValueError`` with "protected" in
+the message; the router turns that into a 403.
+
+``get_model`` re-applies the organization filter by hand on top of the ambient scope filter,
+because a leaked model row hands out a stored provider API key. The other functions go
+through ``crud_utils``, which applies it for them.
+
+The block that moved here also held ``test_model_connection``, a stub that never did
+anything but return ``True`` -- the real check was never written. It had no callers (the
+``/models/{id}/test`` endpoint uses ``ModelConnectionService`` instead), so it was dropped
+rather than carried over.
+"""
+
+import uuid
+from typing import List, Optional
+
+from sqlalchemy.orm import Session
+
+from rhesis.backend.app import models, schemas
+from rhesis.backend.app.utils.crud_utils import (
+    create_item,
+    delete_item,
+    get_items_detail,
+    update_item,
+)
+from rhesis.backend.app.utils.query_utils import include
+
+
+def get_model(
+    db: Session, model_id: uuid.UUID, organization_id: str = None, user_id: str = None
+) -> Optional[models.Model]:
+    """Get a specific model by ID with its related objects and organization filtering"""
+    query = (
+        db.query(models.Model)
+        .options(include(models.Model.provider_type))
+        .filter(models.Model.id == model_id)
+    )
+
+    # Apply organization filtering (SECURITY CRITICAL)
+    if organization_id:
+        from uuid import UUID as UUIDType
+
+        query = query.filter(models.Model.organization_id == UUIDType(organization_id))
+
+    return query.first()
+
+
+def get_models(
+    db: Session,
+    skip: int = 0,
+    limit: int = 10,
+    sort_by: str = "created_at",
+    sort_order: str = "desc",
+    filter: str | None = None,
+    organization_id: str = None,
+    user_id: str = None,
+) -> List[models.Model]:
+    """Get all models with their related objects"""
+    return get_items_detail(
+        db,
+        models.Model,
+        skip,
+        limit,
+        sort_by,
+        sort_order,
+        filter,
+        organization_id=organization_id,
+        user_id=user_id,
+    )
+
+
+def create_model(
+    db: Session, model: schemas.ModelCreate, organization_id: str = None, user_id: str = None
+) -> models.Model:
+    """Create a new model."""
+    return create_item(db, models.Model, model, organization_id, user_id)
+
+
+def update_model(
+    db: Session,
+    model_id: uuid.UUID,
+    model: schemas.ModelUpdate,
+    organization_id: str = None,
+    user_id: str = None,
+) -> Optional[models.Model]:
+    """Update a model."""
+    # First check if the model is protected
+    existing_model = get_model(db, model_id, organization_id)
+    if existing_model and getattr(existing_model, "is_protected", False):
+        # For protected models, only allow updating certain fields
+        # (tags, comments, status, owner, assignee)
+        # Block updates to core model configuration properties
+        protected_fields = {
+            "name",
+            "model_name",
+            "provider_type_id",
+            "key",
+            "endpoint",
+            "is_protected",
+            "icon",
+        }
+
+        # Convert model to dict and check if any protected fields are being updated
+        update_data = (
+            model.model_dump(exclude_unset=True)
+            if hasattr(model, "model_dump")
+            else model.dict(exclude_unset=True)
+        )
+
+        # Check if user is trying to change any protected fields to a different value
+        attempted_protected_updates = []
+        for field in protected_fields:
+            if field in update_data:
+                existing_value = getattr(existing_model, field)
+                new_value = update_data[field]
+                # Only flag as error if the value is actually changing
+                if existing_value != new_value:
+                    attempted_protected_updates.append(field)
+
+        if attempted_protected_updates:
+            fields_str = ", ".join(attempted_protected_updates)
+            raise ValueError(
+                f"Cannot update protected fields ({fields_str}) on system model. "
+                "Only tags, status, owner, and assignee can be modified."
+            )
+
+    return update_item(db, models.Model, model_id, model, organization_id, user_id)
+
+
+def delete_model(
+    db: Session, model_id: uuid.UUID, organization_id: str, user_id: str
+) -> Optional[models.Model]:
+    """Delete a model (protected models cannot be deleted)"""
+    # First check if the model is protected
+    model = get_model(db, model_id, organization_id)
+    if model and getattr(model, "is_protected", False):
+        raise ValueError("Cannot delete protected system model")
+
+    return delete_item(db, models.Model, model_id, organization_id=organization_id, user_id=user_id)
