+++ tests/backend/security/test_organization_filtering.py
@@ -11,6 +11,7 @@ import pytest
 from sqlalchemy.orm import Session
 
 from rhesis.backend.app import crud, models
+from rhesis.backend.app.crud import model as model_crud
 from rhesis.backend.app.crud import task as task_crud
 from rhesis.backend.app.crud.metric import create_metric, get_metric
 from rhesis.backend.app.crud.test_run import get_test_run
@@ -420,12 +421,12 @@ class TestCrudOrganizationFiltering:
             endpoint="https://test.example.com",
             key="test-key",
         )
-        model = crud.create_model(
+        model = model_crud.create_model(
             db=test_db, model=model_data, organization_id=str(org1.id), user_id=str(user1.id)
         )
 
         # User from org1 should be able to access the model
-        result_org1 = crud.get_model(
+        result_org1 = model_crud.get_model(
             test_db, model.id, organization_id=str(org1.id), user_id=str(user1.id)
         )
         assert result_org1 is not None
@@ -433,7 +434,7 @@ class TestCrudOrganizationFiltering:
         assert result_org1.organization_id == org1.id
 
         # User from org2 should NOT be able to access the model
-        result_org2 = crud.get_model(
+        result_org2 = model_crud.get_model(
             test_db, model.id, organization_id=str(org2.id), user_id=str(user2.id)
         )
         assert result_org2 is None
@@ -612,7 +613,7 @@ class TestCrudParameterValidation:
             ("get_test_run", get_test_run),
             ("get_endpoint", crud.get_endpoint),
             ("get_prompt", crud.get_prompt),
-            ("get_model", crud.get_model),
+            ("get_model", model_crud.get_model),
             ("get_metric", get_metric),
         ]
